Racas

+++ src/com/timsoft/meurebanho/model/Fazenda.java
@@ -2,7 +2,7 @@ package com.timsoft.meurebanho.model;
 
 public class Fazenda extends IDDescricao{
 
-	public Fazenda(long id, String descricao) {
+	public Fazenda(int id, String descricao) {
 		super(id, descricao);
 	}
 }
