+++ src/com/timsoft/meurebanho/model/Lote.java
@@ -2,7 +2,7 @@ package com.timsoft.meurebanho.model;
 
 public class Lote extends IDDescricao{
 
-	public Lote(long id, String descricao) {
+	public Lote(int id, String descricao) {
 		super(id, descricao);
 	}
 }
