+++ src/com/timsoft/meurebanho/model/Pasto.java
@@ -2,7 +2,7 @@ package com.timsoft.meurebanho.model;
 
 public class Pasto extends IDDescricao {
 
-	public Pasto(long id, String descricao) {
+	public Pasto(int id, String descricao) {
 		super(id, descricao);
 	}
 }
